UI: Improve typings for CardLayout and CardActions components
Change-type: patch
