+++ apps/ui/lib/lens/full/SingleCard/SingleCard.tsx
@@ -54,7 +54,7 @@ export default class SingleCardFull extends React.Component<any, any> {
 	}
 
 	render() {
-		const { card, channel, types, actionItems } = this.props;
+		const { card, channel, types } = this.props;
 
 		const type = helpers.getType(card.type, types);
 
@@ -67,12 +67,7 @@ export default class SingleCardFull extends React.Component<any, any> {
 		const displayTimeline = card.type !== 'user';
 
 		return (
-			<CardLayout
-				overflowY
-				card={card}
-				channel={channel}
-				actionItems={actionItems}
-			>
+			<CardLayout overflowY card={card} channel={channel}>
 				<Divider width="100%" color={helpers.colorHash(card.type)} />
 
 				<SingleCardTabs
